refactor: use integrations/userServiceClient.js for all user lookups; remove services/userDirectory.js; pass auth header for real external data
