+++ controllers/staff.controller.js
@@ -1,6 +1,6 @@
 const { Staff } = require('../models/userModels');
 const { crudController } = require('./basic.crud');
-const { listStaff, getStaffById } = require('../services/userDirectory');
+const { listStaff, getStaffById } = require('../integrations/userServiceClient');
 
 const base = { ...crudController(Staff) };
